Rollback to previous deployed version
This reverts commit 21ede116f1db5e435e366011558f5d08952dd139.
