+++ helm_deploy/values-preprod.yaml
@@ -140,7 +140,6 @@ generic-service:
       - circleci
       - internal
       - prisons # do not remove, the prisons group also covers probation
-      - probation
 
   scheduledDowntime:
     enabled: true
