+++ helm_deploy/values-prod.yaml
@@ -139,7 +139,6 @@ generic-service:
     groups:
       - internal
       - prisons # do not remove, the prisons group also covers probation
-      - probation
 
 generic-prometheus-alerts:
   alertSeverity: make-recall-decision-prod
