+++ server/views/pages/recommendations/responseToProbation.njk
@@ -1,14 +1,18 @@
 {% extends "../../partials/layout.njk" %}
+
 {% set pageTitle = makePageTitle({ pageHeading: pageTitles[page.id], hasErrors: errors }) %}
+
 {% block beforeContent %}
     {{ backLink() }}
 {% endblock %}
+
 {% block content %}
     <div class="govuk-grid-row">
         <div class="govuk-grid-column-two-thirds-from-desktop">
             {% include '../../partials/error-summary.njk' %}
             <form novalidate method="post">
-                <input type="hidden" name="_csrf" value="{{ csrfToken }}"/> {{ govukTextarea({
+                <input type="hidden" name="_csrf" value="{{ csrfToken }}" />
+                {{ govukTextarea({
                     name: "responseToProbation",
                     id: "responseToProbation",
                     label: {
@@ -17,7 +21,7 @@
                         isPageHeading: true
                     },
                     hint: {
-                        html: "<p class='govuk-body'>Give details of overall compliance. This includes good behaviour.</p><p class='govuk-body'>Your answer will be recorded in NDelius.</p>"
+                        html: "<p class='govuk-body'>Give details of overall compliance. This includes good behaviour. Your answer will go into a Part A form if you decide to recall " + recommendation.personOnProbation.name + ".</p><p class='govuk-body'>" + recommendation.personOnProbation.name + " will see the Part A.</p>"
                     },
                     value: inputDisplayValues.value,
                     errorMessage: errorMessage(errors.responseToProbation)
@@ -27,3 +31,4 @@
         </div>
     </div>
 {% endblock %}
+
